add confirmation for delete and amend comments
