+++ public/index.html
@@ -25,6 +25,7 @@
     <noscript>
       You need to enable JavaScript to run this app.
     </noscript>
+    
     <div id="root"></div>
     <!--
       This HTML file is a template.
