--- src/Header.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const Header = () => (
-  <div>
-    <span>
-      <h2>Contacts</h2>
-      <button id="add-contact-button" class="btn btn-primary" onClick={this.handleClick} >Add Contact</button>
-    </span>
-  </div>
-)
-
-export default Header
