--- src/Schedule.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Schedule = () => (
-  <h1>Schedule</h1>
-)
-
-export default Schedule
