+++ src/components/Contacts.js
@@ -18,12 +18,14 @@ class Contacts extends React.Component {
     this.toggleRedirect = this.toggleRedirect.bind(this);
   }
 
+  // handle button click to add contact - sets redirect prop to true (cf. <Route path='/contacts' below)
   handleButtonClick = () => {
     this.setState(() => ({
       redirect: true
     }))
   }
 
+  // pass this prop to the ContactsNew component to enable re-initializing redirect prop
   toggleRedirect = () => {
     this.setState(() => ({
       redirect: false
