+++ src/index.js
@@ -2,8 +2,8 @@ import { BrowserRouter } from 'react-router-dom'
 import React from 'react';
 import ReactDOM from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.css';
-import './index.css';
-import App from './App';
+import './styles/index.css';
+import App from './components/App';
